API, Core: Move all stats-related classes into core and make them package-private
This moves all stats related code into iceberg-core to avoid any potential API breakages before the Spec has been finalized.
It also moves all classes under the org.apache.iceberg package for usability/visibility in other classes v4-related classes.
